Library: remove explicit nil termination in list implementation
While it is true that a proper list is a chain of pairs of which the
last one holds nil in its cdr, in this implentation lists are
represented not by pairs but by F# lists. The nil termination would
actually be in the last car of the list. Removing it allows to take
advantage of the F# list primitives.