Library: implement a first Eval function
Eval evalutates a S-expression in a scope. Only primitives & special
forms are supported. Functions definition and macro expansion are not.
Eval uses two stacks to evaluate S-expressions as per Bel evaluation
rules (left to right, depth first, car of sexpr is a function, a
primitive or a special form, values bound to symbols). An instruction
stack and a data stack.
To evaluate a S-Expression, the instruction stack is initialised with
it. The evaluation proceed until the instruction stack is emptied. The
result is the last value of the data stack.

Example: evalutating (+ 1 2)
EvalSexpr (+ 1 2) ->
  Match Sexpr
  PushData (Sexpr 1)
  PushData (Sexpr 2)
  PushInstr (EvalTop 2)
  PushInstr (EvalSexpr +)
EvalSexpr + ->
  match Symbol
  PushData (lookup +)
EvalTop 2 ->
  Match Primitive
  PushInstr (EvalPrimitive PopData)
  PushInstr (EvalSexpr PopData)
  PushInstr (EvalSexpr PopData)
EvalSexpr 1 ->
  Match Number
  PushData 1 ->
EvalSexpr 2
  Match Number
  PushData 2
EvalPrimitive add 2
  add 2
    PopData
    PopData
    PushData 3

3 = PopData