Library/Lisp: move Primitive, SpecialForm & Error under Atom; remove Value
Primitive, SpecialForm & Error are atoms though they are not
readable. Moving these to Atom allows to use Sexpr as a ubuquitous
value representation. That means less type nesting and to be able to
manipulate primitives and special forms as data (for instance allow to
get them from the environment).
The main drawback is that equality can no longer be derived for Atoms
and Sexpr because of function signatures.