Library/Evaluator: implement literal function call
A literal function, or user-defined function, is defined as follow:
(lit clo environment parameters body)
environment is an alist of bound symbols (usually nil or scope).
parameters is a list of symbols.
body is a single S-expression.

When a literal function is in the car of a list to be evaluated, each element
of the cdr of the list is evaluated from left to right, bound to the parameters
then the body is evaluated in this environment.

> ((lit clo nil (x y) (+ x y)) 1 2)
6