Library/Parser: implement the parser using FParsec
FParsec is a parser combinator library. The benefits of using it are
mainly simplicity, conciseness and better error messages (better than
non existent). For instance, adding support for characters or quoted
strings should be simple.
I am not sure about the performance aspect of this particular
implementation. Each time a list must be parsed, the parser tries to
parse a pair until the dot. This might be an issue (or not).