Library/SpecialForms: implement set
Bound a symbol to a value. Set evaluates its arguments so to set the
symbol appearing in code it must be quoted. Set tries to set the symbol
in the dynamic scope if it already exists, then in the lexical scope
if it already exists and finally in the global environment if existing
or not.