Library/SpecialForms: allow lit to refer to any primitive or form
Since symbols in the global environment can be rebound and lexical and
dynamic environments have precedence over it at lookup, a way to access
the primitives and special forms must be provided.
Calling (lit prim _) or (lit form _) provides that feature. To be always
available, lit is implemented as a special case in lookup so redefining
it has no effect.